refactor: docs-at-cursor into DocsCommand in docs.js
