+++ lib/kite.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { DEFAULT_MAX_FILE_SIZE } = require('kite-api');
+const { DocsCommand } = require('./docs');
 
 // Contents of this plugin will be reset by Kite on start.
 // Changes you make are not guaranteed to persist.
@@ -18,7 +19,6 @@ let child_process,
   RollbarReporter,
   OverlayManager,
   KiteEditors,
-  DataLoader,
   KSG,
   KSGShortcut,
   DisposableEvent,
@@ -99,7 +99,6 @@ const Kite = (module.exports = {
       NotificationsCenter = require('./notifications-center');
       RollbarReporter = require('./rollbar-reporter');
       metrics = require('./metrics.js');
-      DataLoader = require('./data-loader');
       KiteEditors = require('./editors');
       Status = require('./status');
       KSG = require('./ksg');
@@ -265,11 +264,12 @@ const Kite = (module.exports = {
           OverlayManager.dismiss();
           this.closeKSG();
         },
-        'kite:docs-at-cursor': () => {
-          this.expandAtCursor();
-        },
       })
     );
+
+    const docsCmd = new DocsCommand();
+    docsCmd.add();
+    this.subscriptions.add(docsCmd);
 
     this.subscriptions.add(
       atom.commands.add('atom-overlay kite-expand', {
@@ -372,52 +372,6 @@ const Kite = (module.exports = {
         pane.setActiveItem(item);
       }
     }
-  },
-
-  expandAtCursor(editor) {
-    editor = editor || atom.workspace.getActiveTextEditor();
-
-    if (!editor) {
-      return;
-    }
-
-    const position = editor.getLastCursor().getBufferPosition();
-
-    this.highlightWordAtPosition(editor, position);
-
-    if (!editor.getPath()) {
-      return;
-    }
-
-    DataLoader.getHoverDataAtPosition(editor, position).then(data => {
-      const [symbol] = data.symbol;
-
-      if (symbol && symbol.id) {
-        atom.applicationDelegate.openExternal(`kite://docs/${symbol.id}`);
-      }
-    });
-  },
-
-  highlightWordAtPosition(editor, position, cls = '') {
-    if (!VirtualCursor) {
-      VirtualCursor = require('./virtual-cursor');
-    }
-
-    const cursor = new VirtualCursor(editor, position);
-    const range = cursor.getCurrentWordBufferRange({
-      includeNonWordCharacters: false,
-    });
-    const marker = editor.markBufferRange(range, {
-      invalidate: 'touch',
-    });
-    const decoration = editor.decorateMarker(marker, {
-      type: 'highlight',
-      class: `expand-at-cursor-highlight ${cls}`,
-      item: this,
-    });
-
-    // Timed for all transition to be finished by then
-    setTimeout(() => decoration.destroy(), 800);
   },
 
   getStatusItem() {
